fix write function bug
